modify response data

+++ routes/post.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const { Post } = require("../models");
-const getPaging = require("../utils/getPaging.js");
+const getHasNext = require("../utils/getHasNext.js");
 const upload = require("../middleware/upload.js");
 const s3 = require("../aws.js");
 
@@ -12,12 +12,11 @@ router.get("/", async (req, res) => {
     if (!req.sessionId) throw new Error("Invalid Session");
     const { cursor, limit } = req.query;
     const posts = await Post.find(cursor ? { _id: { $lt: cursor } } : {})
-      .populate({ path: "writer", select: "_id name photoUrl" })
-      .select("-thumbnail -comments")
+      .populate({ path: "writer", select: "id name photoUrl" })
       .sort({ _id: -1 })
       .limit(limit);
-    const paging = await getPaging(Post, cursor, limit);
-    res.send({ message: "success", posts, paging });
+    const hasNext = await getHasNext(Post, cursor, limit);
+    res.send({ posts, hasNext });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
@@ -28,14 +27,14 @@ router.get("/:id", async (req, res) => {
     if (!req.sessionId) throw new Error("Invalid Session");
     const { id } = req.params;
     const post = await Post.findById(id)
-      .populate({ path: "writer", select: "_id name photoUrl" })
+      .populate({ path: "writer", select: "id name photoUrl" })
       .populate({
         path: "comments",
         select: "-post",
-        populate: { path: "writer", select: "_id name photoUrl" },
+        populate: { path: "writer", select: "id name photoUrl" },
         options: { limit: 10 },
       });
-    res.send({ message: "success", post });
+    res.send({ post });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
@@ -55,7 +54,8 @@ router.post("/", upload.array("image"), async (req, res) => {
       images,
       writer: req.userId,
     });
-    res.send({ message: "success", post });
+    await Post.populate(post, { path: "writer", select: "id name photoUrl" });
+    res.send({ post });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
@@ -71,34 +71,7 @@ router.delete("/:id", async (req, res) => {
         return s3.deleteObject({ Bucket, Key: image.key }).promise();
       })
     );
-    res.send({ message: "success" });
-  } catch (error) {
-    res.status(400).send({ message: error.message });
-  }
-});
-
-router.patch("/:id", async (req, res, next) => {
-  try {
-    if (!req.sessionId) throw new Error("Invalid Session");
-    const { id } = req.params;
-    const { title, contents, deletedKeys } = req.body;
-
-    const post = await Post.findByIdAndUpdate(
-      id,
-      { title, contents },
-      { new: true }
-    );
-
-    if (deletedKeys.length < 1) {
-      res.send({ message: "success" });
-    } else {
-      req.body = {
-        id,
-        deletedKeys,
-        images: post.images,
-      };
-      next();
-    }
+    res.send({ id });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
@@ -106,21 +79,26 @@ router.patch("/:id", async (req, res, next) => {
 
 router.patch("/:id", async (req, res) => {
   try {
-    let { id, deletedKeys, images } = req.body;
+    if (!req.sessionId) throw new Error("Invalid Session");
+    const { id } = req.params;
+    const { title, contents, deletedKeys } = req.body;
 
-    deletedKeys.forEach((deletedKey) => {
-      images = images.filter((image) => image.key !== deletedKey);
-    });
+    let post = await Post.findById(id);
+    post.title = title;
+    post.contents = contents;
 
-    await Post.findByIdAndUpdate(id, { images });
-
-    await Promise.all(
-      deletedKeys.map((deletedKey) =>
-        s3.deleteObject({ Bucket, Key: deletedKey }).promise()
-      )
-    );
-
-    res.send({ message: "success" });
+    if (deletedKeys.length > 1) {
+      await Promise.all(
+        deletedKeys.map((deletedKey) =>
+          s3.deleteObject({ Bucket, Key: deletedKey }).promise()
+        )
+      );
+      deletedKeys.forEach((deletedKey) => {
+        post.images = post.images.filter((image) => image.key !== deletedKey);
+      });
+    }
+    post = await post.save();
+    res.send({ post });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
@@ -136,7 +114,7 @@ router.patch("/:id/like", async (req, res) => {
       { $inc: { likeCount: isLike ? 1 : -1 } },
       { new: true }
     );
-    res.send({ message: "success", post });
+    res.send({ post });
   } catch (error) {
     res.status(400).send({ message: error.message });
   }
